Add item refund flow and adjust item qty types
Implement partial per-item refunds and related schema/model updates.

- Add refundItem API (TransactionController): validates input, checks item ownership/status, restores menu stock, computes refund amount, decrements transaction totals, marks item as 'Refund', and cancels the transaction if all items are refunded; all inside a DB transaction with error handling.
- Register new route POST /transactions/{id}/refund-item in routes/api.php.
- Update TransactionItem model: include user_id in fillable and cast qty as double to support fractional quantities.
- Update transaction_items migration: change qty column to double to allow decimal quantities.
- Update settings migration: include user_id, keep store and receipt fields, set tax_rate/service_charge as decimal, add service_mode default 'fast_casual', and add foreign key constraint to users (multi-tenant readiness).

These changes enable safe partial refunds, fractional item quantities (e.g. 0.5), and multi-user/store settings.

+++ app/Models/TransactionItem.php
@@ -11,10 +11,9 @@ class TransactionItem extends Model
 
     protected $table = 'transaction_items';
 
-    // [PENTING] Masukkan user_id disini juga
     protected $fillable = [
         'transaction_id',
-        'user_id',      // <--- WAJIB ADA (Untuk filter di Dapur/Laporan per item)
+        'user_id',      
         'menu_name',
         'qty',
         'price',
@@ -22,12 +21,12 @@ class TransactionItem extends Model
         'status'
     ];
 
+    // [UBAH INI] Pastikan qty di-cast sebagai double
     protected $casts = [
-        'qty' => 'integer',
+        'qty' => 'double', // <--- FIX: Sekarang Laravel akan mempertahankan angka desimal (0.5)
         'price' => 'double',
     ];
 
-    // Relasi Balik ke Transaksi
     public function transaction()
     {
         return $this->belongsTo(Transaction::class, 'transaction_id', 'id');
